feat(expense edit): implements expense editing functionality
Implements the feature to edit and update expenses.
When the user clicks on the 'Edit' button for a specific expense in the expense table, the form used to register an expense enters edit mode, and the fields are filled in with the data of the selected expense to be edited.